Extract curves & text functions
